All patches by Zack Middleton: - Bug 5057  - Two weapon number range checks allows invalid number - Bug 5058  - Railgun explosion cannot be colorized - Bug 5059  - Client's railgun does not glow in third person and mirror - Bug 5060  - Warning in RAVENMD4 R_GetAnimTag cross compiling with amd64-mingw32

+++ code/renderer/tr_model.c
@@ -1160,7 +1160,7 @@ void R_GetAnimTag( mdrHeader_t *mod, int framenum, const char *tagName, md3Tag_t
 
 			// uncompressed model...
 			//
-			frameSize = (long)( &((mdrFrame_t *)0)->bones[ mod->numBones ] );
+			frameSize = (intptr_t)( &((mdrFrame_t *)0)->bones[ mod->numBones ] );
 			frame = (mdrFrame_t *)((byte *)mod + mod->ofsFrames + framenum * frameSize );
 
 			for (j = 0; j < 3; j++)
